requirements/base.txt
gunicorn==21.2.0
Django==4.2.6
django-filter==23.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
Markdown==3.5
django-environ==0.11.2
Pillow==10.2.0
django-cors-headers==4.3.0
celery==5.3.6
django-celery-beat==2.5.0
django-celery-results==2.5.1
redis==5.0.1
drf-yasg==1.21.7
kavenegar==1.1.2 # kavenegar sms service provider
django-extensions==3.2.3
ipython==8.20.0
psycopg2-binary==2.9.6
django-taggit==5.0.1
django-js-asset==2.2.0
django-ckeditor==6.6.1
# -------------------------

